requirements.txt
asgiref==3.2.7
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cloudinary==1.20.0
dj-database-url==0.5.0
Django==3.0.3
django-appconf==1.0.4
django-cloudinary-storage==0.2.3
django-cors-headers==3.2.1
django-filter==2.2.0
django-googledrive-storage==1.5.0
django-heroku==0.3.1
djangorestframework==3.11.0
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.14.3
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
gunicorn==20.0.4
httplib2==0.17.3
idna==2.9
mock==4.0.2
oauth2client==4.1.3
Pillow==7.0.0
protobuf==3.12.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.0.1
